primitif matematik hesap yapilir halde

+++ src/Main.java
@@ -10,7 +10,7 @@ public class Main {
                 continue;
             }
             Menu.printOutput();
-            int result = CommandParser.execute(parsedCommand);
+            double result = CommandParser.execute(parsedCommand);
             System.out.println(result);
         } while (!command.equals(Constants.COMMAND_EXIT));
         Menu.exit();
